Slå sammen context builders og test-oppsamlere mellom modell og app
