+++ modell/build.gradle.kts
@@ -2,7 +2,7 @@ dependencies {
     api(project(":kontrakter-intern"))
     api(project(":kontrakter-ekstern-v2"))
     api(project(":kontrakter-frontend:dtoer"))
-    api(project(":testdata"))
+    testImplementation(project(":testdata"))
     api(project(":felles"))
     api("com.fasterxml.jackson.core:jackson-databind:2.21.3")
     implementation("io.ktor:ktor-http:3.5.0")
