--- src/test/kotlin/no/nav/tilbakekreving/endring/EndringObservatørOppsamler.kt
@@ -1,45 +1,0 @@
-package no.nav.tilbakekreving.endring
-
-import no.nav.tilbakekreving.fagsystem.Ytelse
-import no.nav.tilbakekreving.kontrakter.behandling.Behandlingsstatus
-import no.nav.tilbakekreving.kontrakter.beregning.Vedtaksresultat
-import no.nav.tilbakekreving.kontrakter.periode.Datoperiode
-import no.nav.tilbakekreving.kontrakter.tilstand.TilbakekrevingTilstand
-import java.math.BigDecimal
-import java.time.OffsetDateTime
-import java.util.UUID
-
-class EndringObservatørOppsamler : EndringObservatør {
-    override fun behandlingsstatusOppdatert(
-        behandlingId: UUID,
-        forrigeBehandlingId: UUID?,
-        eksternFagsystemId: String,
-        eksternBehandlingId: String,
-        ytelse: Ytelse,
-        tilstand: TilbakekrevingTilstand,
-        behandlingstatus: Behandlingsstatus,
-        vedtaksresultat: Vedtaksresultat?,
-        venterPåBruker: Boolean,
-        ansvarligEnhet: String?,
-        ansvarligSaksbehandler: String,
-        ansvarligBeslutter: String?,
-        totaltFeilutbetaltBeløp: BigDecimal?,
-        totalFeilutbetaltPeriode: Datoperiode?,
-    ) {}
-
-    override fun behandlingEndret(behandlingEndret: EndringObservatør.BehandlingEndret) {}
-
-    override fun vedtakFattet(
-        behandlingId: UUID,
-        forrigeBehandlingId: UUID?,
-        behandlingOpprettet: OffsetDateTime,
-        eksternFagsystemId: String,
-        eksternBehandlingId: String,
-        ytelse: Ytelse,
-        vedtakFattetTidspunkt: OffsetDateTime,
-        ansvarligEnhet: String?,
-        ansvarligSaksbehandler: String,
-        ansvarligBeslutter: String,
-        vurderteUtbetalinger: List<VurdertUtbetaling>,
-    ) {}
-}
