+++ testdata/build.gradle.kts
@@ -10,6 +10,7 @@ repositories {
 }
 
 dependencies {
+    api(project(":modell"))
     testImplementation(platform("org.junit:junit-bom:6.1.0"))
     testImplementation("org.junit.jupiter:junit-jupiter")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
